spltting stream and streamProps
