+++ src/index.ts
@@ -3,6 +3,8 @@ export { scanPlans } from "./observable/scanPlans"
 export { fromPlan } from "./operators/fromPlan"
 export { patchScan } from "./operators/patchScan"
 export { spreadMap } from "./operators/spreadMap"
+export { tapWhen } from "./operators/tapWhen"
 export { toPlan } from "./operators/toPlan"
 export { plan } from "./plan"
-export { Stream, stream, StreamProps, streamProps } from "./stream"
+export { Stream, stream } from "./stream"
+export { StreamProps, streamProps } from "./streamProps"
